feat(bank): split to test categories and helpers
