+++ packages/clarity/unit-tests/taral-bank.base.test.ts
@@ -1,5 +1,6 @@
 import { Cl } from "@stacks/transactions";
 import { describe, expect, it } from "vitest";
+import { expectUsdaTransfer } from "./helpers/transfer";
 
 const accounts = simnet.getAccounts();
 const WALLET_1 = accounts.get("wallet_1")!;
@@ -212,161 +213,5 @@ describe("Taral bank test flows", () => {
     const events = acceptFinancingResult.events.filter((event: any) => event.event === 'ft_transfer_event');
     expectUsdaTransfer(events[0].data, DEPLOYER, WALLET_2, borrow - downPayment);
     expectUsdaTransfer(events[1].data, DEPLOYER, WALLET_2, downPayment);
-    }),
-
-    it("Should not be able to place a financing offer after another one has been accepted", () => {
-        const financingId = 1;
-
-        const purchaseOrderResult = simnet.callPublicFn(
-            "taral-bank",
-            "create-purchase-order",
-            [
-                Cl.uint(borrow),
-                Cl.uint(downPayment),
-                Cl.standardPrincipal(WALLET_2) // the seller
-            ], WALLET_1
-        );
-
-        expect(purchaseOrderResult.result).toBeOk(Cl.uint(purchaseOrderId));
-        expectUsdaTransfer(purchaseOrderResult.events[0].data, WALLET_1, DEPLOYER, downPayment);
-
-        // place a financing offer
-        let placeFinancingResult = simnet.callPublicFn(
-            "taral-bank",
-            "finance",
-            [
-                Cl.uint(purchaseOrderId),
-                Cl.bool(false)
-            ], WALLET_3
-        );
-
-        // console.log(JSON.stringify(placeFinancingResult, null, 2));
-        expect(placeFinancingResult.result).toBeOk(Cl.uint(financingId)); // financing id is 1
-        expectUsdaTransfer(placeFinancingResult.events[0].data, WALLET_3, DEPLOYER, borrow - downPayment);
-
-        const acceptFinancingResult = simnet.callPublicFn(
-            "taral-bank",
-            "accept-financing",
-            [
-                Cl.uint(1),
-            ], WALLET_1
-        );
-
-        expect(acceptFinancingResult.result).toBeOk(Cl.uint(financingId));
-        const events = acceptFinancingResult.events.filter((event: any) => event.event === 'ft_transfer_event');
-        expectUsdaTransfer(events[0].data, DEPLOYER, WALLET_2, borrow - downPayment);
-        expectUsdaTransfer(events[1].data, DEPLOYER, WALLET_2, downPayment);
-
-        // place a financing offer
-        placeFinancingResult = simnet.callPublicFn(
-            "taral-bank",
-            "finance",
-            [
-                Cl.uint(purchaseOrderId),
-                Cl.bool(false)
-            ], WALLET_3
-        );
-
-        expect(placeFinancingResult.result).toBeErr(Cl.uint(122));
-    }),
-
-    it("Should be able to check if the loan has defaulted", () => {
-        const financingId = 1;
-
-        const purchaseOrderResult = simnet.callPublicFn(
-            "taral-bank",
-            "create-purchase-order",
-            [
-                Cl.uint(borrow),
-                Cl.uint(downPayment),
-                Cl.standardPrincipal(WALLET_2) // the seller
-            ], WALLET_1
-        );
-
-        expect(purchaseOrderResult.result).toBeOk(Cl.uint(purchaseOrderId));
-        expectUsdaTransfer(purchaseOrderResult.events[0].data, WALLET_1, DEPLOYER, downPayment);
-
-        // place a financing offer
-        let placeFinancingResult = simnet.callPublicFn(
-            "taral-bank",
-            "finance",
-            [
-                Cl.uint(purchaseOrderId),
-                Cl.bool(false)
-            ], WALLET_3
-        );
-
-        // console.log(JSON.stringify(placeFinancingResult, null, 2));
-        expect(placeFinancingResult.result).toBeOk(Cl.uint(financingId)); // financing id is 1
-        expectUsdaTransfer(placeFinancingResult.events[0].data, WALLET_3, DEPLOYER, borrow - downPayment);
-
-        const acceptFinancingResult = simnet.callPublicFn(
-            "taral-bank",
-            "accept-financing",
-            [
-                Cl.uint(1),
-            ], WALLET_1
-        );
-
-        expect(acceptFinancingResult.result).toBeOk(Cl.uint(financingId));
-        const events = acceptFinancingResult.events.filter((event: any) => event.event === 'ft_transfer_event');
-        expectUsdaTransfer(events[0].data, DEPLOYER, WALLET_2, borrow - downPayment);
-        expectUsdaTransfer(events[1].data, DEPLOYER, WALLET_2, downPayment);
-
-        let hasPoDefaulted = simnet.callReadOnlyFn(
-            "taral-bank",
-            "is-po-defaulted",
-            [
-                Cl.uint(purchaseOrderId),
-            ], WALLET_1
-        );
-
-        expect(hasPoDefaulted.result).toBeOk(Cl.bool(false));
-
-        fastForwardMonths(1);
-
-        hasPoDefaulted = simnet.callReadOnlyFn(
-            "taral-bank",
-            "is-po-defaulted",
-            [
-                Cl.uint(purchaseOrderId),
-            ], WALLET_1
-        );
-
-        expect(hasPoDefaulted.result).toBeOk(Cl.bool(true));
     })
-
-    function fastForwardMonths(months: number) {
-        const blocksPerMonth = 6 * 24 * 31; // 6 blocks per hour, 24 hours per day, 31 days per month
-        const blocksToFastForward = months * blocksPerMonth;
-
-        for (let i = 0; i < blocksToFastForward; i++) {
-            simnet.mineEmptyBlock();
-        }
-    }
-
-    /*
-     * Helper function to assert that a transfer event is a USDa transfer 
-    */
-    function expectUsdaTransfer(transferEvent: any, sender: string, recipient: string, amount: any) {
-
-        let senderAddress = sender;
-        if (sender === DEPLOYER) {
-            senderAddress = `${sender}.taral-bank`;
-        }
-
-        let recipientAddress = recipient;
-        if (recipient === DEPLOYER) {
-            recipientAddress = `${recipient}.taral-bank`;
-        }
-
-
-        expect(transferEvent.asset_identifier).toStrictEqual(
-            `${DEPLOYER}.usda-token::usda`,
-        );
-
-        expect(transferEvent.sender).toStrictEqual(senderAddress);
-        expect(transferEvent.recipient).toStrictEqual(recipientAddress);
-        expect(transferEvent.amount).toStrictEqual(`${amount}`);
-    }
 });
